+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `encrypt` setting to model definitions that when enabled will encrypt the property value in the database.
 - Added polymorphic relationship type.
 - Saved values are no longer validated by `Model::valid()`. Only unsaved values will be validated.
+- Callable validation rules can now set their own validation error message.
 
 ### Changed
 - Make model internal properties private when possible.